Moved fixture sdk_node_created_after_some_txns to top level
